+++ lib/src/router.dart
@@ -1,10 +1,7 @@
 // Define a function to handle named routes in order to support
 // Flutter web url navigation and deep linking.
 import 'package:curie/src/views/error/error_view.dart';
-import 'package:curie/src/views/navigation_test/first_widget.dart';
 import 'package:curie/src/views/navigation_test/navigation_test_view.dart';
-import 'package:curie/src/views/navigation_test/second_widget.dart';
-import 'package:curie/src/views/navigation_test/third_widget.dart';
 import 'package:curie/src/views/route_test/route_test_view.dart';
 import 'package:curie/src/views/settings/settings_view.dart';
 import 'package:flutter/material.dart';
@@ -14,31 +11,6 @@ import 'package:go_router/go_router.dart';
 final GoRouter router = GoRouter(
   initialLocation: '/',
   routes: <RouteBase>[
-    ShellRoute(
-      // builder: (context, state, child) {
-      //   return child;
-      // },
-      routes: <RouteBase>[
-        GoRoute(
-          path: FirstWidget.routeName,
-          builder: (BuildContext context, GoRouterState state) {
-            return const FirstWidget();
-          },
-        ),
-        GoRoute(
-          path: SecondWidget.routeName,
-          builder: (BuildContext context, GoRouterState state) {
-            return const FirstWidget();
-          },
-        ),
-        GoRoute(
-          path: ThirdWidget.routeName,
-          builder: (BuildContext context, GoRouterState state) {
-            return const FirstWidget();
-          },
-        ),
-      ],
-    ),
     GoRoute(
       path: NavigationTestView.routeName,
       builder: (BuildContext context, GoRouterState state) {